+++ tools/AtrImage.h
@@ -39,18 +39,18 @@ public:
 	void DetermineDiskFormatFromLayout();
 	void CalculateImageSize();
 
-	unsigned int GetSectorLength(unsigned int sector) const;
+	uint16_t GetSectorLength(uint16_t sector) const;
 
 	EDiskFormat fDiskFormat;
 
 	ESectorLength fSectorLength;
 
-	unsigned int fSectorsPerTrack;
-	unsigned int fTracksPerSide;
-	unsigned int fSides;
+	uint16_t fSectorsPerTrack;
+	uint8_t fTracksPerSide;
+	uint8_t fSides;
 
-	unsigned int fNumberOfSectors;
-	unsigned int fImageSize;
+	uint16_t fNumberOfSectors;
+	uint32_t fImageSize;
 };
 
 class AtrImage : public DiskImage {
@@ -61,30 +61,30 @@ public:
 	virtual ~AtrImage();
 
 	virtual ESectorLength GetSectorLength() const { return fImageConfig.fSectorLength; }
-	virtual unsigned int GetSectorLength(unsigned int sectorNumber) const { return fImageConfig.GetSectorLength(sectorNumber); }
-	virtual unsigned int GetNumberOfSectors() const { return fImageConfig.fNumberOfSectors; }
+	virtual uint16_t GetSectorLength(uint16_t sectorNumber) const { return fImageConfig.GetSectorLength(sectorNumber); }
+	virtual uint16_t GetNumberOfSectors() const { return fImageConfig.fNumberOfSectors; }
 
 	const AtrImageConfig& GetImageConfig() { return fImageConfig; }
 
 	EDiskFormat GetDiskFormat() const { return fImageConfig.fDiskFormat; }
 
-	unsigned int GetSectorsPerTrack() const { return fImageConfig.fSectorsPerTrack; }
-	unsigned int GetTracksPerSide() const { return fImageConfig.fTracksPerSide; }
-	unsigned int GetSides() const { return fImageConfig.fSides; }
+	uint16_t GetSectorsPerTrack() const { return fImageConfig.fSectorsPerTrack; }
+	uint8_t GetTracksPerSide() const { return fImageConfig.fTracksPerSide; }
+	uint8_t GetSides() const { return fImageConfig.fSides; }
 
-	virtual unsigned int GetImageSize() const { return fImageConfig.fImageSize; }
+	virtual size_t GetImageSize() const { return fImageConfig.fImageSize; }
 
-	virtual bool ReadSector(unsigned int sector,
+	virtual bool ReadSector(uint16_t sector,
 		       uint8_t* buffer,
-		       unsigned int buffer_length) const;
+		       size_t buffer_length) const;
 
-	virtual bool WriteSector(unsigned int sector,
+	virtual bool WriteSector(uint16_t sector,
 		       const uint8_t* buffer,
-		       unsigned int buffer_length);
+		       size_t buffer_length);
 
 	virtual bool CreateImage(EDiskFormat format) = 0;
-	virtual bool CreateImage(ESectorLength density, unsigned int sectors) = 0;
-	virtual bool CreateImage(ESectorLength density, unsigned int sectorsPerTrack, unsigned int tracks, unsigned int sides) = 0;
+	virtual bool CreateImage(ESectorLength density, uint16_t sectors) = 0;
+	virtual bool CreateImage(ESectorLength density, uint16_t sectorsPerTrack, uint8_t tracks, uint8_t sides) = 0;
 
 	virtual bool IsAtrImage() const;
 	virtual bool IsAtrMemoryImage() const;
@@ -94,13 +94,14 @@ public:
 
 protected:
 	bool SetFormat(EDiskFormat format);
-	bool SetFormat(ESectorLength density, unsigned int numberOfSectors);
-	bool SetFormat(ESectorLength density, unsigned int sectors, unsigned int tracks, unsigned int sides);
+	// note: only 1..65535 sectors are allowed
+	bool SetFormat(ESectorLength density, uint32_t numberOfSectors);
+	bool SetFormat(ESectorLength density, uint16_t sectors, uint8_t tracks, uint8_t sides);
 
 	bool SetFormatFromATRHeader(const uint8_t* header);
 	bool CreateATRHeaderFromFormat(uint8_t* header) const;
 
-	int CalculateOffset(unsigned int sector) const;
+	ssize_t CalculateOffset(uint16_t sector) const;
 	// -1 = error
 
 private:
@@ -110,7 +111,7 @@ private:
 	AtrImageConfig fImageConfig;
 };
 
-inline int AtrImage::CalculateOffset(unsigned int sector) const
+inline ssize_t AtrImage::CalculateOffset(uint16_t sector) const
 {
 	if ( (sector == 0) || (sector > fImageConfig.fNumberOfSectors) ) {
 		return -1;
@@ -180,7 +181,7 @@ inline bool AtrImageConfig::operator!=(const AtrImageConfig& other) const
 inline AtrImageConfig::~AtrImageConfig()
 { }
 
-inline unsigned int AtrImageConfig::GetSectorLength(unsigned int sector) const
+inline uint16_t AtrImageConfig::GetSectorLength(uint16_t sector) const
 {
 	if (sector == 0 || sector > fNumberOfSectors) {
 		return 0;
